+++ jest.setup.js
@@ -3,4 +3,6 @@ jest.mock('./src/utils/stdin')
 
 require('css.escape') // Polyfill for CSS.escape
 
+CSS.registerProperty = jest.fn()
+
 process.env.FORCE_COLOR = '0'